+++ test/acceptance/attributes.js
@@ -1,0 +1,162 @@
+// FLUSHALL Redis before starting
+
+var   assert        = require('../support/assert')
+    , tests         = module.exports = {}
+    , _             = require('underscore')
+    , querystring   = require('querystring')
+    , fs            = require('fs')
+    , redis         = require('redis')
+    , th            = require('../support/test_helper')
+    , Step          = require('step')
+    , mapnik        = require('mapnik')
+    , Windshaft     = require('../../lib/windshaft')
+    , ServerOptions = require('../support/server_options')
+    , http          = require('http');
+
+suite('attributes', function() {
+
+    ////////////////////////////////////////////////////////////////////
+    //
+    // SETUP
+    //
+    ////////////////////////////////////////////////////////////////////
+
+    var server = new Windshaft.Server(ServerOptions);
+    server.setMaxListeners(0);
+    var redis_client = redis.createClient(ServerOptions.redis.port);
+
+    checkCORSHeaders = function(res) {
+      var h = res.headers['access-control-allow-headers'];
+      assert.ok(h);
+      assert.equal(h, 'X-Requested-With, X-Prototype-Version, X-CSRF-Token');
+      var h = res.headers['access-control-allow-origin'];
+      assert.ok(h);
+      assert.equal(h, '*');
+    };
+
+    suiteSetup(function(done) {
+
+      // Check that we start with an empty redis db 
+      redis_client.keys("*", function(err, matches) {
+          if ( err ) { done(err); return; }
+          assert.equal(matches.length, 0, "redis keys present at setup time:\n" + matches.join("\n"));
+          done();
+      });
+
+    });
+
+    test("can only be fetched from layer having an attributes spec",
+    function(done) {
+
+      var mapconfig =  {
+        version: '1.1.0',
+        layers: [
+           { type: 'mapnik', options: {
+               sql: "select 1 as id, 'SRID=4326;POINT(0 0)'::geometry as the_geom",
+               cartocss: '#style { }',
+               cartocss_version: '2.0.1'
+             } },
+           { type: 'mapnik', options: {
+               sql: "select 1 as i, 6 as n, 'SRID=4326;POINT(0 0)'::geometry as the_geom",
+               attributes: { id:'i', columns: ['n'] },
+               cartocss: '#style { }',
+               cartocss_version: '2.0.1'
+             } }
+        ]
+      };
+      var expected_token; 
+      Step(
+        function do_post()
+        {
+          var next = this;
+          assert.response(server, {
+              url: '/database/windshaft_test/layergroup',
+              method: 'POST',
+              headers: {'Content-Type': 'application/json' },
+              data: JSON.stringify(mapconfig)
+          }, {}, function(res, err) { next(err, res); });
+        },
+        function checkPost(err, res) {
+          if ( err ) throw err;
+          assert.equal(res.statusCode, 200, res.statusCode + ': ' + res.body);
+          // CORS headers should be sent with response
+          // from layergroup creation via POST
+          checkCORSHeaders(res);
+          var parsedBody = JSON.parse(res.body);
+          if ( expected_token ) assert.deepEqual(parsedBody, {layergroupid: expected_token, layercount: 2});
+          else expected_token = parsedBody.layergroupid;
+          return null;
+        },
+        function do_get_attr_0(err)
+        {
+          if ( err ) throw err;
+          var next = this;
+          assert.response(server, {
+              url: '/database/windshaft_test/layergroup/' + expected_token + '/0/attributes/1',
+              method: 'GET'
+          }, {}, function(res, err) { next(err, res); });
+        },
+        function check_error_0(err, res) {
+          if ( err ) throw err;
+          assert.equal(res.statusCode, 400, res.statusCode + ( res.statusCode != 200 ? (': ' + res.body) : '' ));
+          var parsed = JSON.parse(res.body);
+          assert.equal(parsed.error, "Layer 0 has no exposed attributes");
+          return null;
+        },
+        function do_get_attr_1(err)
+        {
+          if ( err ) throw err;
+          var next = this;
+          assert.response(server, {
+              url: '/database/windshaft_test/layergroup/' + expected_token + '/1/attributes/1',
+              method: 'GET'
+          }, {}, function(res, err) { next(err, res); });
+        },
+        function check_attr_1(err, res) {
+          if ( err ) throw err;
+          assert.equal(res.statusCode, 200, res.statusCode + ': ' + res.body);
+          var parsed = JSON.parse(res.body);
+          assert.deepEqual(parsed, {"n":6});
+          return null;
+        },
+        function finish(err) {
+          var errors = [];
+          if ( err ) errors.push(''+err);
+          redis_client.exists("map_cfg|" +  expected_token, function(err, exists) {
+              if ( err ) errors.push(err.message);
+              //assert.ok(exists, "Missing expected token " + expected_token + " from redis");
+              redis_client.del("map_cfg|" +  expected_token, function(err) {
+                if ( err ) errors.push(err.message);
+                if ( errors.length ) done(new Error(errors));
+                else done(null);
+              });
+          });
+        }
+      );
+    });
+
+    ////////////////////////////////////////////////////////////////////
+    //
+    // TEARDOWN
+    //
+    ////////////////////////////////////////////////////////////////////
+
+    suiteTeardown(function(done) {
+
+      // Check that we left the redis db empty
+      redis_client.keys("*", function(err, matches) {
+          try {
+            assert.equal(matches.length, 0, "Left over redis keys:\n" + matches.join("\n"));
+          } catch (err2) {
+            if ( err ) err.message += '\n' + err2.message;
+            else err = err2;
+          }
+          redis_client.flushall(function() {
+            done(err);
+          });
+      });
+
+    });
+
+});
+